Fix !visualprospectingnearspawn (#30)
* Update BS

* Update dependencies.gradle

* Update HooksShared.java

Just a reformat + comment

* Add final patches + spotless

This SHOULD fix VPNearSpawn. Also generates a ton of random edits because spotless said so ig.

* Update dependencies.gradle

Add the electric scanner mod for debugging.

* Add recache command, remove auto recache

I can't replicate the automatic bug or fix in a full environment, so I'm removing that code for now. Instead, a command to only recache spawn has been added.

* spotless + BS

* Update dependencies.gradle

* Update dependencies.gradle

* Apply simple suggestions from SinTh0r4s

Co-authored-by: Johann Bernhardt <johann.bernhardt@posteo.de>

* Apply complex suggestions from SinTh0r4s

* Update build.gradle

* Update deps

---------

Co-authored-by: Johann Bernhardt <johann.bernhardt@posteo.de>
Co-authored-by: Martin Robertz <dream-master@gmx.net>